etl/extract/requirements.txt
python-dotenv
requests
fuzzywuzzy
python-Levenshtein
pprinter
inquirer